Take the GIL in the new image callback
The Global Interpreter Lock (GIL) *must* be acquired by non-Python
threads before using *any* of the Python API, so ensure this is done in
the image callback function, which will not be in the main Python
thread.

Since the callback has the GIL, it is no longer necessary for the images
queue to have its own mutex.  All accesses to the queue will be
serialized by the GIL.