Add SVSAsyncError
Add a new exception, SVSAsyncError, which is used to indicate a problem
in the new image callback, since a normal exception will not be handled
appropriately.

This exception is raised in the main thread from the callback with
raise_asynchronous_exception().

This exception is not very descriptive; it has no description, and the
traceback can be misleading, but the only other option is to completely
ignore all errors in the callback.